# artifacts/api-server/src/routes/drivers.ts
import { Router, type IRouter } from "express";
import { eq, and, desc } from "drizzle-orm";
import { db, driversTable, usersTable, ordersTable, orderItemsTable, walletTransactionsTable } from "@workspace/db";
import { RegisterDriverBody, UpdateDriverStatusBody, UpdateDriverLocationBody } from "@workspace/api-zod";
import { CASH_LIMIT } from "../lib/dispatch";

const router: IRouter = Router();

function formatDriver(d: typeof driversTable.$inferSelect, user?: typeof usersTable.$inferSelect | null) {
  return {
    id: d.id,
    userId: d.userId,
    isOnline: d.isOnline,
    isLocked: d.isLocked,
    rating: d.rating,
    acceptanceRate: d.acceptanceRate,
    currentLat: d.currentLat,
    currentLng: d.currentLng,
    cashBalance: d.cashBalance,
    walletBalance: d.walletBalance,
    totalDeliveries: d.totalDeliveries,
    createdAt: d.createdAt,
    user: user ? {
      id: user.id,
      name: user.name,
      email: user.email,
      phone: user.phone,
      role: user.role,
      isBanned: user.isBanned,
      createdAt: user.createdAt,
    } : undefined,
  };
}

router.post("/drivers/register", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const parsed = RegisterDriverBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
  const existing = await db.select().from(driversTable).where(eq(driversTable.userId, sessionUserId));
  if (existing.length > 0) { res.status(409).json({ error: "Already registered as driver" }); return; }
  const [driver] = await db.insert(driversTable).values({
    userId: sessionUserId,
    vehicleType: parsed.data.vehicleType,
    vehiclePlate: parsed.data.vehiclePlate ?? null,
  }).returning();
  const [user] = await db.select().from(usersTable).where(eq(usersTable.id, sessionUserId));
  res.status(201).json(formatDriver(driver, user));
});

router.get("/drivers/me", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, sessionUserId));
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
  const [user] = await db.select().from(usersTable).where(eq(usersTable.id, sessionUserId));
  res.json(formatDriver(driver, user));
});

router.patch("/drivers/me/status", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const parsed = UpdateDriverStatusBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, sessionUserId));
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
  if (driver.isLocked && parsed.data.isOnline) {
    res.status(403).json({ error: "Driver is locked due to cash limit. Please settle your balance." });
    return;
  }
  const [updated] = await db.update(driversTable).set({ isOnline: parsed.data.isOnline }).where(eq(driversTable.userId, sessionUserId)).returning();
  const [user] = await db.select().from(usersTable).where(eq(usersTable.id, sessionUserId));
  res.json(formatDriver(updated, user));
});

router.patch("/drivers/me/location", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const parsed = UpdateDriverLocationBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
  await db.update(driversTable).set({ currentLat: parsed.data.lat, currentLng: parsed.data.lng }).where(eq(driversTable.userId, sessionUserId));
  res.json({ success: true });
});

router.get("/drivers/me/wallet", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, sessionUserId));
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }

  const today = new Date();
  today.setHours(0, 0, 0, 0);
  const weekAgo = new Date();
  weekAgo.setDate(weekAgo.getDate() - 7);
  weekAgo.setHours(0, 0, 0, 0);

  const allTxns = await db.select().from(walletTransactionsTable).where(eq(walletTransactionsTable.driverId, driver.id));
  const todayTxns = allTxns.filter(t => new Date(t.createdAt) >= today);
  const weekTxns = allTxns.filter(t => new Date(t.createdAt) >= weekAgo);

  const earningsToday = todayTxns.filter(t => t.type === "earning" || t.type === "bonus").reduce((s, t) => s + t.amount, 0);
  const earningsWeek = weekTxns.filter(t => t.type === "earning" || t.type === "bonus").reduce((s, t) => s + t.amount, 0);

  const todayOrders = await db.select().from(ordersTable)
    .where(and(eq(ordersTable.driverId, driver.id), eq(ordersTable.status, "delivered")));
  const deliveriesToday = todayOrders.filter(o => new Date(o.createdAt) >= today).length;
  const deliveriesWeek = todayOrders.filter(o => new Date(o.createdAt) >= weekAgo).length;

  const nextMilestone = Math.ceil((driver.totalDeliveries + 1) / 10) * 10;
  const progress = ((driver.totalDeliveries % 10) / 10) * 100;

  res.json({
    walletBalance: driver.walletBalance,
    cashBalance: driver.cashBalance,
    totalEarningsToday: earningsToday,
    totalEarningsWeek: earningsWeek,
    deliveriesToday,
    deliveriesWeek,
    pendingBonus: nextMilestone >= 20 ? 800 : 300,
    cashLimitReached: driver.cashBalance >= CASH_LIMIT,
  });
});

router.get("/drivers/me/wallet/transactions", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, sessionUserId));
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
  const txns = await db.select().from(walletTransactionsTable)
    .where(eq(walletTransactionsTable.driverId, driver.id))
    .orderBy(desc(walletTransactionsTable.createdAt));
  res.json(txns);
});

router.get("/drivers/available-jobs", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const pendingOrders = await db.select().from(ordersTable).where(eq(ordersTable.status, "pending"));
  res.json(pendingOrders.map(o => ({
    id: o.id,
    customerId: o.customerId,
    businessId: o.businessId,
    driverId: o.driverId,
    status: o.status,
    totalAmount: o.totalAmount,
    deliveryFee: o.deliveryFee,
    commission: o.commission,
    driverEarnings: o.driverEarnings,
    paymentMethod: o.paymentMethod,
    isPaid: o.isPaid,
    deliveryAddress: o.deliveryAddress,
    notes: o.notes,
    createdAt: o.createdAt,
    items: [],
  })));
});

router.post("/drivers/jobs/:orderId/accept", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  const raw = Array.isArray(req.params.orderId) ? req.params.orderId[0] : req.params.orderId;
  const orderId = parseInt(raw, 10);
  if (isNaN(orderId)) { res.status(400).json({ error: "Invalid orderId" }); return; }
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, sessionUserId));
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
  if (driver.isLocked) { res.status(403).json({ error: "Driver is locked" }); return; }
  const [order] = await db.update(ordersTable).set({ driverId: driver.id, status: "accepted" }).where(and(eq(ordersTable.id, orderId), eq(ordersTable.status, "pending"))).returning();
  if (!order) { res.status(409).json({ error: "Order no longer available" }); return; }
  const allItems = await db.select().from(orderItemsTable).where(eq(orderItemsTable.orderId, orderId));
  res.json({
    id: order.id,
    customerId: order.customerId,
    businessId: order.businessId,
    driverId: order.driverId,
    status: order.status,
    totalAmount: order.totalAmount,
    deliveryFee: order.deliveryFee,
    commission: order.commission,
    driverEarnings: order.driverEarnings,
    paymentMethod: order.paymentMethod,
    isPaid: order.isPaid,
    deliveryAddress: order.deliveryAddress,
    notes: order.notes,
    createdAt: order.createdAt,
    items: allItems,
  });
});

router.post("/drivers/jobs/:orderId/decline", async (req, res): Promise<void> => {
  const sessionUserId = (req.session as any)?.userId;
  if (!sessionUserId) { res.status(401).json({ error: "Unauthorized" }); return; }
  res.json({ success: true });
});

export default router;