# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Que Lo Que delivery platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers/register:
    post:
      operationId: registerDriver
      tags: [drivers]
      summary: Register as a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDriverBody"
      responses:
        "201":
          description: Driver registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
...
  /drivers/me/status:
    patch:
      operationId: updateDriverStatus
      tags: [drivers]
      summary: Toggle driver online/offline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverStatusBody"
      responses:
        "200":
          description: Driver status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/me/location:
    patch:
      operationId: updateDriverLocation
      tags: [drivers]
      summary: Update driver GPS location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverLocationBody"
      responses:
        "200":
          description: Location updated
...
  /drivers/jobs/{orderId}/accept:
    post:
      operationId: acceptJob
      tags: [drivers]
      summary: Accept a delivery job
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /drivers/jobs/{orderId}/decline:
    post:
      operationId: declineJob
      tags: [drivers]
      summary: Decline a delivery job
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job declined

  # ADMIN
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, driver, business, admin]
        isBanned:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isBanned
        - createdAt
...
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [food, supermarket, pharmacy, liquor]
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        isActive:
          type: boolean
        rating:
          type: number
        totalOrders:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - category
        - address
        - isActive
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        businessId:
          type: integer
        driverId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, picked_up, delivered, cancelled]
        totalAmount:
          type: number
        deliveryFee:
          type: number
        commission:
          type: number
        driverEarnings:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card]
        isPaid:
          type: boolean
        deliveryAddress:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        business:
          $ref: "#/components/schemas/Business"
        customer:
          $ref: "#/components/schemas/User"
        driver:
          $ref: "#/components/schemas/Driver"
        customerRating:
          type: integer
        driverRating:
          type: integer
        businessRating:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - customerId
        - businessId
        - status
        - totalAmount
        - deliveryFee
        - paymentMethod
        - isPaid
        - deliveryAddress
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - orderId
        - productName
        - quantity
        - price
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        isOnline:
          type: boolean
        isLocked:
          type: boolean
        rating:
          type: number
        acceptanceRate:
          type: number
        currentLat:
          type: number
        currentLng:
          type: number
        cashBalance:
          type: number
        walletBalance:
          type: number
        totalDeliveries:
          type: integer
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - isOnline
        - isLocked
        - rating
        - acceptanceRate
        - cashBalance
        - walletBalance
        - totalDeliveries
        - createdAt
...
    RegisterDriverBody:
      type: object
      properties:
        vehicleType:
          type: string
        vehiclePlate:
          type: string
      required:
        - vehicleType
    UpdateDriverStatusBody:
      type: object
      properties:
        isOnline:
          type: boolean
      required:
        - isOnline
    UpdateDriverLocationBody:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required:
        - lat
        - lng
...